+++ js_modules/dagster-ui/packages/ui-core/src/assets/AssetView.tsx
@@ -12,7 +12,7 @@ import {ASSET_NODE_DEFINITION_FRAGMENT, AssetNodeDefinition} from './AssetNodeDe
 import {ASSET_NODE_INSTIGATORS_FRAGMENT} from './AssetNodeInstigatorTag';
 import {AssetNodeLineage} from './AssetNodeLineage';
 import {AssetNodeOverview, AssetNodeOverviewNonSDA} from './AssetNodeOverview';
-import {AssetPageHeader} from './AssetPageHeader';
+import {AssetPageHeader} from './AssetPageHeader.oss';
 import {AssetPartitions} from './AssetPartitions';
 import {AssetPlotsPage} from './AssetPlotsPage';
 import {AssetTabs} from './AssetTabs';
