+++ js_modules/dagster-ui/packages/ui-core/src/assets/AssetsGroupsGlobalGraphRoot.tsx
@@ -7,7 +7,7 @@ import {
   globalAssetGraphPathFromString,
   globalAssetGraphPathToString,
 } from './globalAssetGraphPathToString';
-import {useAssetDefinitionFilterState} from './useAssetDefinitionFilterState';
+import {useAssetDefinitionFilterState} from './useAssetDefinitionFilterState.oss';
 import {useTrackPageView} from '../app/analytics';
 import {AssetGraphExplorer} from '../asset-graph/AssetGraphExplorer';
 import {AssetGraphFetchScope} from '../asset-graph/useAssetGraphData';
