+++ js_modules/dagster-ui/packages/ui-core/src/assets/AssetsOverviewRoot.tsx
@@ -5,7 +5,7 @@ import {Box} from '@dagster-io/ui-components';
 import React, {useMemo} from 'react';
 import {useHistory, useParams} from 'react-router-dom';
 
-import {AssetGlobalLineageLink, AssetPageHeader} from './AssetPageHeader';
+import {AssetGlobalLineageLink, AssetPageHeader} from './AssetPageHeader.oss';
 import {AssetView} from './AssetView';
 import {AssetsCatalogTable} from './AssetsCatalogTable';
 import {assetDetailsPathForKey} from './assetDetailsPathForKey';
