+++ js_modules/dagster-ui/packages/ui-core/src/assets/useAssetCatalogFiltering.oss.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 
 import {useAssetGroupSelectorsForAssets} from './AssetGroupSuggest';
 import {AssetTableFragment} from './types/AssetTableFragment.types';
-import {useAssetDefinitionFilterState} from './useAssetDefinitionFilterState';
+import {useAssetDefinitionFilterState} from './useAssetDefinitionFilterState.oss';
 import {useAssetSearch} from './useAssetSearch';
 import {CloudOSSContext} from '../app/CloudOSSContext';
 import {isCanonicalStorageKindTag} from '../graph/KindTags';
